Client can run flexsearch as a treat
